make `SSEStreamHandler` generic
