+++ Sources/GPTBridge/Internal/SSE Handling/StreamingRequestManager.swift
